+++ agents/critic.py
@@ -1,9 +1,17 @@
 from abc import abstractmethod
 
-from agents import TorchAgent, State, Value
+from agents.common import TorchAgent, State, Value
 
 
 class BaseCritic(TorchAgent, config_name='base_critic'):
     @abstractmethod
+    def forward(self, state: State) -> Value:
+        pass
+
+
+class TowerCritic(BaseCritic, config_name='tower_critic'):
+    def __init__(self):
+        super().__init__()
+
     def forward(self, state: State) -> Value:
         pass
