+++ main.py
@@ -1,5 +1,4 @@
-from agents import BaseAgent
-from agents.ppo import PPOAgent
+from agents.common import BaseAgent
 from utils import parse_args
 
 
@@ -9,7 +8,7 @@ def train():
     # BaseAgent.create_from_config(params['agent_1'])  # random
     # BaseAgent.create_from_config(params['agent_2'])  # dqn
     # BaseAgent.create_from_config(params['agent_3'])  # reinforce
-    PPOAgent.create_from_config(params['agent_4'])  # ppo
+    BaseAgent.create_from_config(params['agent_4'])  # ppo
 
 
 if __name__ == '__main__':
